runner.cli parsers set as a library (#5021)
* INIT 1/2

* INIT 2/2

* LINT

---------
